+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
  - `fold_change_mode` parameter for heatmap to only show fold-change instead of split tiles
  - implement `RunInfo` utility to fetch information about previous runs
  - bugfix passing `ax` to `kompot.plot.embedding`
+ - implemented `mgroups` in `kompot.plot.embedding` to plot multiple groupings
 
 ## [0.3.3]